Changes to avoid gpinitsystem failure when multiple address interfaces and mirrors. (#13049) (#13063)
This regression was introduced while fixing bash banner issue 248c1c740b2f5f351ef6ee6cc4c1317397c857f9
Due to this output of ifaddrs command is interpreted as text instead of array which in turn
resulted in some ip missing in pg_hba.conf. That led to failure of gpinitsystem when multiple
interfaces are present and cluster created with mirrors.