Revert "Fix index corruption when invalid snapshot used with AO tables. (#12797)"
This reverts commit 44dc41a212b98b7b60c470f8f85ffc72894526fe.

This introduced a compile error on pipeline when -Werror=maybe-uninitialized
is given. We somehow missed the pipeline check in the original PR.

The fix should be simple but since this is blocking release 6.19.3 now and
it takes time to verify the fix on pipeline, we'll just revert it and
re-submit a fixed version of the commit.