Fix segno being incorrectly reset in mdsync()
`segno` in `mdsync()` is reset to 0 if `is_seg0` is true.
That is not correct because `is_seg0` really just means
something's wrong when opening the base relfilenode in
`_mdfd_getseg()`, which could happen for any `segno`.
So, `segno` shouldn't be changed, especially considering
`bms_next_member()` relies on it to give a correct
next `segno`.

Also limit the scope of `is_seg0` so its value won't
affect things unintentionally in the following loops.

The corrupted `segno` value could result in stuck loop
in `mdsync()`. We've seen cases where relfilenodes of
some table get unlinked while we are still trying to
access `segno=1` in `mdsync()`. After getting ENOENT
(file not existing) for the file, we will reset `segno`
to 0. Then because the `canceled` flag was set we'll
break the retry loop and start accessing `segno=1` again,
and everything repeats.