Fix flaky unit test in buildMirrorSegments
While testing the recover_mirrors/add_mirrors function, we assert that
we make the right call to the gpsegrecovery wrapper. This unit test
failed on the 6X pipeline because for some reason gplog's verbose
setting got enabled. Because of this, the call to gpsegreovery.py
included a '-v' flag. This commit explicitly mocks the function
`logging_is_verbose` and sets the return value to False.