Adds --continue-check-on-fatal option for check flag
- Need to be used along with check flag `-c`
- Indent log messages for proper output rendering

The --continue-check-on-fatal flag lets us traverse through the
pg_upgrade --check list without exiting the first error encountered. It
allows the user to get a report of the entire list of non-upgradable
objects in their cluster at one go.

Note: We can't ignore certain checks even with the flag; they may be
critical to the check process . One such example is
check_proper_datallowconn(), which ensures that we can connect to user
databases to perform the checks themselves.  Specific checks are
multi-step, like the check to ensure that only the install user is the
sole user in the target database. If we get fatal: "could not determine
the number of users", we can't proceed with the check. Thus, in these
cases, the flag will not respected.

- Example:

```
pg_upgrade \
-b /usr/local/5XSTABLE/bin/ \
-B /usr/local/6XSTABLE/bin/ \
-d /data/5XStable/demoDataDir-1 \
-D /data/6XStable/demoDataDir-1 \
-p 15432 \
-P 6000 \
--old-gp-dbid=1 \
--new-gp-dbid=1 \
-c \
--continue-check-on-fatal
```

This is a backport of gpdb master commit:
4308039ef24f74dedea66dc5300d3b7f16489486

Backport changes/conflicts resolved:
- some checks are different in 6X_STABLE
- updating usage for new flags is different
- initialize greenplum user options are different we need to add support
  of new and old gp-dbid

Co-authored-by: Gaurab Dey <gaurabd@vmware.com>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@vmware.com>