fix cache look up failed for relation 25 in pg_catalog.gp_dump_query_oids() (#13283)
This PR is trying to fix issue #13195. 

### RCA
The RCA is that PostgreSQL does not support type conversion between unknown and text,
but Greenplum hacks this behavior so that the below SQL can be executed on 6X:

CREATE TABLE t (pid varchar(1), dummy varchar(1)) DISTRIBUTED BY (pid);
CREATE VIEW s AS SELECT 'D' as dummy;
select * from t join s on t.dummy = s.dummy;	// not allowed in PostgreSQL 9

In the process of hack, the principle is multi-level function nesting, first convert unknown into
cstring through unknownout function, and then use textin function to convert cstring into
text. Then in the nested process, the constructed function parameters are:

```
args = list_make3(fe,
        makeConst(OIDOID, -1, InvalidOid, sizeof(Oid),
	        ObjectIdGetDatum(intypioparam),
	        false, true),
        makeConst(INT4OID, -1, InvalidOid, sizeof(int32),
	        Int32GetDatum(-1),
	        false, true));
```

The problem lies in OIDOID. Greenplum itself uses it as a placeholder, or a default parameter,
because in the actual type conversion execution process, the executor does not care what type
of Const you are. Will only take constvalue.

However, in the implementation of pg_catalog.gp_dump_query_oids(), the entire query tree
will be traversed, the RelationID will be saved and returned. In the fix_expr_common() function.
There is this logic:

```
static void
fix_expr_common(PlannerInfo *root, Node *node)
{
    else if (IsA(node, Const))
    {
          Const	   *con = (Const *) node;
          
          /* Check for regclass reference */
          if (ISREGCLASSCONST(con))
              root->glob->relationOids = lappend_oid(
                    root->glob->relationOids, 
                    DatumGetObjectId(con->constvalue));
    }
```

And OIDOID is a legal REGCLASSCONST. Therefore, ObjectIdGetDatum(intypioparam)
will be thrown into RelationOids, and an error will be reported when it is finally judged whether
it has sub class.

### Work Around Fix
I have to say, The way to fix above problem in this PR is not canonical, and is a little weird, but I
do not have a better idea.

When performing the actual unknown -> text type conversion, such as performing HASH JOIN,
the executor does not care what type of the second parameter of the conversion function is, and
the OID type is added only to satisfy the number and type of function parameters, so this change
does not affect Planner. But ORCA prints some information, and that's the only place affected.