[6X] add original storage options to AO table with index when expanding (#13248)
When a table is an AO table and has an index, ALTER TABLE EXPAND TABLE will not be done directly
using CTAS, but create table + insert data. However, when creating a table, the original storage option
of the AO table is ignored, and the default parameters of the current environment are used, which will
lead to inconsistency between the table before and after expansion.

The pr fixes issue #13245, reconstructs the storage option of the original table, and passes it down.

This is the backport of #13246.