Fix planner bug: correctly handling NULL when pulling-up NOT-IN.
Greenplum will pull up not-in sublink to a specific join LASJ,
this kind of join's joinqual might contain a NULL const here,
for such case we do not need to split it. A case that can
reach here is:

  create table t1(a int not null, b int not null);
  create table t2(a int not null, b int not null);
  explain  select 1 from t1 where (NULL, b) not in (select a, b from t2);

The above SQL in Greenplum will be turned in a join whose qual contains
a bool expr (NULL = t2.a) and (t1.b = t2.b), this piece of expr will be
evaluated to (t1.b = t2.b) and NULL by the following code path:
  subquery_planner
    -> preprocess_qual_conditions(root, (Node *) parse->jointree)
    -> preprocess_expression
    -> eval_const_expressions
    -> eval_const_expressions_mutator

So in splitJoinQualExpr, we should expect a null const in LASJ's
qual's bool expr's arg.

Fix Issue: https://github.com/greenplum-db/gpdb/issues/13212.