Add support in ORCA for splitting ordered-set agg
This commit adds a new pre-processing step in the preprocessor that
splits a query containing an ordered-set agg into gp_percentile agg.
Below is the preporcessed query:

Input query:
```
+--CLogicalGbAgg( Global )
   |--CLogicalGet "t" ("t")
   +--CScalarProjectList
      +--CScalarProjectElement "percentile_cont" (9)
         +--CScalarAggFunc (percentile_cont , Distinct: false , Aggregate Stage: Global)
            |--CScalarValuesList
            |  +--CScalarCast
            |     +--CScalarIdent "a" (0)
            |--CScalarValuesList
            |  +--CScalarConst (0.500)
            |--CScalarValuesList
            |  +--CScalarSortGroupClause(tleSortGroupRef:0,eqop:670,sortop:672,nulls_first:false,hashable:true)
            +--CScalarValuesList
```

Output preprocessed query:
```
Common Table Expressions:
+--CLogicalCTEProducer (0)
   +--CLogicalGet "t" ("t")

Algebrized preprocessed query:
+--CLogicalCTEAnchor (0)
   +--CLogicalGbAgg( Global )
      |--CLogicalLimit ( (97,1.0), "a" (0), NULLsLast )
      |  |--CLogicalNAryJoin
      |  |  |--CLogicalCTEConsumer (0)
      |  |  |--CLogicalGbAgg( Global )
      |  |  |  |--CLogicalCTEConsumer (0)
      |  |  |  +--CScalarProjectList
      |  |  |     +--CScalarProjectElement "ColRef_0029" (29)
      |  |  |        +--CScalarAggFunc (count , Distinct: false , Aggregate Stage: Global)
      |  |  |           |--CScalarValuesList
      |  |  |           |  +--CScalarIdent "a" (19)
      |  |  |           |--CScalarValuesList
      |  |  |           |--CScalarValuesList
      |  |  |           +--CScalarValuesList
      |  |  +--CScalarConst (1)
      |  |--CScalarConst (0)
      |  +--CScalarConst (null)
      +--CScalarProjectList
         +--CScalarProjectElement "percentile_cont" (9)
            +--CScalarAggFunc (gp_percentile , Distinct: false , Aggregate Stage: Global)
               |--CScalarValuesList
               |  |--CScalarCast
               |  |  +--CScalarIdent "a" (0)
               |  |--CScalarConst (0.500)
               |  +--CScalarIdent "ColRef_0029" (29)
               |--CScalarValuesList
               |--CScalarValuesList
               +--CScalarValuesList
```

This commit also adds/updates the following to support the
the newly added pre-processing step:

- Add new gpcontrib for calculating percentile on ordered input

Currently we have percentile_cont and percentile_disc that calculates
the continuous and discrete percentile on a column, by first sorting
them in memory. This limits ORCA and planner to generate a plan such that
all the tuples are first Gather'ed on coordinator and then the aggregate
is executed. The transition functions called by these aggregates first
sorts the tuples in memory which may cause slowness when they are
called on multiple columns. To avoid this, we can parallelize the sort
on the segments and then run the aggregate.
This commit adds new contrib with aggregate functions that take in as
input the column, fraction(for percentile), the total count of rows in
the table(excluding NULLs) and distinct count of each and then
calculates the percentile(assumes data is ordered). Each percentile
aggregate has a corresponding new aggregate function.
These functions will be used in the following commits by ORCA to generate
a plan that split the ordered-set aggregates into gp_percentile aggs.
Also added explain tests.

- Update GetAggregate to pass in the number of args

Currently `GetAggregate()` calls the gpdb function `get_aggregate()` which is
only used by ORCA (no other refs to it). Passing in number of args as
the newly added gp_percentile_agg's for which we want to fetch the OID
have >1 args.

- Add utilities used to support splitting an ordered-agg

Currently we only support the following cases when we want to split an
ordered-set agg into its corresponding gp_percentile agg:
 - The sort expression doesn't have any computed columns
 - The fraction is not a const ARRAY
 - No GROUP BY clause
 - Only inbuilt ordered-set aggs are supported (percentile_cont &
   percentile_disc)

- Add support for passing corresponding gp_percentile_agg

Since the corresponding gp_percentile_agg that we wish to split the
inbuilt ordered aggs into is created as an extension, we need to pass
along the OID of it that will be used by ORCA's preprocessor to split
it. Since this is only passed if the corresponding extension exists,
creating it as an optional param.