pg_dump: Rework binary upgrade metadata queries.
Backported from be8bb094d2ae8ec24c03a399611507a5f1a8f0d9

Changes made:
* To gather binary upgrade information for partition children,
  they need to exist in the tblinfo array for findTableByOid.
  This requires the getTable query to also collect all
  partition children, then filter them out of the dump so we
  only dump the partition root DDL. Previously, partition children
  were being excluded from the getTables query. Now they are gathered
  and not dumped.
* AOCO tables do have toast tables in GPDB6, so do not filter them out
  when setting toast oids.
* Additional minor changes to getTypes and getTables were made to fit
  the commit into the framework of 6X_STABLE pg_dump since
  those functions have repetitive SQL which has been refactored
  upstream.
* Removed force_array_type from binary_upgrade_set_type_oids_by_type_oid.
  This argument is only relevant when upgrading from a pre v11 server to
  post v11.

Original commit message follows:

This commit refactors the functions used to gather metadata required for binary
upgrade schema dumps to avoid repetitive queries in performance-critical paths.

Upgrades require us to preassign the OIDs of tables, indexes, and types so that
they match in the old and new clusters. Previously, multiple queries were
executed for each object to gather the necessary info to record in the dump
file. Now we gather all of the info required up front and use hash table lookups
to set and dump the correct OIDs.

The names of pg_toast, pg_aoseg, and pg_bitmapindex tables for the given
table/index are set to match what they will be in the new cluster, using the OID
of the owning table. In many cases these will be the same as the ones in the old
cluster, but not always. Some operations can cause the old table names not to
match its owner's OID, but the new cluster will be using the correct name, and
it's the new cluster's name that we have to use

The AOTableInfo and BMIndxInfo structs have been added, and metadata fields have
been added to the end of the TableInfo, IndxInfo, and TypeInfo structs.

A dump of the regression database using `--binary-upgrade` shows a performance
increase of over 20x and continues to scale as object count increases.

Authored-by: Brent Doil <bdoil@vmware.com>