Avoid further planning in case of result relation was pruned.
A result relation is a target relation for DML operations
in the appropriate queries. If the result relation is a member
of the join tree and is excluded by query predicates, such query
won't modify any tuples. So we could avoid spending resources
for planning remaining join tree relations, especially complex
sub queries. This gives a nice performance gain when planning
DML queries to tables with many leaf tables in the inheritance tree
(i.e. partitioned too). Because Postgres builds a separate plan for
each child relation first according to inheritance_planner function
logic.

There are several iterations to refactor this logic in the upstream
Postgres:

1. Plan Update/Delete query as SELECT query first to collect
list of affected children [1].
2. Rework Modify Table node to handle tuples from any child
relation [2]

So upstream Postgres and gpdb master is not interested in this patch.