Wrong subplans order in correlated subquery.
This problem occurs with Postgres Query Optimizer only, ORCA is not affected.

If we have a partitioned table and a nested correlated subquery, we can face an error
that would say something like that:

FATAL:  Unexpected internal error (list.c:395)
DETAIL:  FailedAssertion("!(n < list->length)", File: "list.c", Line: 395)
In a case when a query has a partitioned table and InitPlan can't be created,
GPDB will make copies of first level subplans by number of partitions, and
all more nested subplans will be added later. But all more nested subplans copies
will be added in a wrong order and that they will go in a flat plan list after thier
parent nodes, what will lead to a crash.

create table t1 (a int) partition by range (a) (start (1) end (3) every (1));
create table t2 (b int8) distributed by (b);
set optimizer = off;
explain select 1 from t1 where a <= (
    SELECT 1 FROM t2
    WHERE t2.b <= (SELECT 1 from t2 as t3 where t3.b = t2.b)
    and t1.a = t2.b);

The intention of this patch is to create nested subplans right on the place. So,
by the time we get in nested_subplan_mutator, the plan already has dedicated
subplans for each partition. The intention of this part of
adjust_appendrel_attrs_mutator is to create a copy of the
most-closely-nested subplan and add it to subplans list. It's implied that
nested subplans of a copy will be added later in fixup_subplans, but it's the
cause of a crush: more nested subplans will be added after their parent subplans
in a flat list. The intension of this fix is to recursively traverse through
all the subplans and add them right on the spot. We make subplan copies
in nested_subplan_mutator and expect that when we get to fixup_subplans,
all the nested subplans are already initialized and fixup_subplans
doens't need to make any copies anymore.