Fix performance issue of syslogger_write_str()  (#13726)
There is a performance issue of logger code: In syslogger_write_str(), 
fwrite() (or write()) was called per char instead of per string so it  
might be called for tens of millions times when the log string is very 
big. The solution is to call fwrite() to write a batch of chars as many
as possible.