add zstd for gpdb6 (#13933)
This pull request adds the zstd file external table support for gpdb6. With this change, gpdb6 can create a readable or writable external table based on files compressed by the zstd algorithm.

The added code defines the functions for reading from a zstd file, which means we can get the content of zstd files through the gpfdist protocol.

Moreover, we also defined the functions for writing into a zstd file, which helps us insert messages in a given format into a zstd file.

With the help of the above functions, gpdb6 can support reading from a zstd file external table, and writing the data into a zstd file external table.

I have added some regression tests for the zstd external table(readable and writable) to make sure the new feature works as expected.