Improve the performance of source_bytes computation during gpexpand.
During gpexpand, it will record all the tables that need to do data
redistribution and in order to show the progress the table size is
also saved in the status_details table.

Previously, we directly invoke pg_relation_size to compute that value
in the target list of a SQL involving only catalog tables. This will
lead to very bad performance because for each table it will do a dispatch
to compute its size and the whole progress is done in serial.

This commit rewrite the SQL generating the status_details information.
The skills used to MPP compute table size is:
  1. we use gp_dist_random('pg_class') to get a force random RangeTable
  2. since pg_relations_size is volatile, wrap the computation in a
     subquery before group-by to prevent optimizer to pull up subquery
     and force the pg_relation_size is evaluated before motion then
     sum it together
  3. join with the main SQL

We add test cases in regress/subselect_gp and regress/db_size_functions
to guard the above SQL skill.

We have done local test to show the performance gain. In a local 8*8
64-segments cluster, we create 100 partition tables, each partition
table contains 2 middle level partitions, each middle level partition
contains 365 leaf partitions. Then in this database, we compare the
speed of the old populate SQL (evaluate pg_relation_size on QD via
dispatching, a tuple a dispatch) and the new SQL (a single SQL and only
need to dispatch once). This patch's SQL only takes about 3.8 sec, however,
the old SQL takes about 50 sec.

Authored-by: Miao Chen <miaochen@mail.ustc.edu.cn>
Authored-by: Zhenghua Lyu <kainwen@gmail.com>