Fix init file usage for gpdb_pitr tests
In a recent commit, the gpdb_pitr test suite added another init file
without the realization that there was an existing one already. This
caused the tests to fail when Orca was enabled. To fix the issue,
simply combine the two init files. We use the newer init file name to
match with GPDB master (7X+).

GPDB 6X commit reference:
https://github.com/greenplum-db/gpdb/commit/4614936f0ef50c9b1f14b2f6ca85aeb4dcfc6d30