pg_upgrade: Remove converter plugin facility.
Backported from 6f56b41ac0cd7092b91834da02e70514d28b8edb

Original commit message follows:

We've not found a use for this so far, and the current need, which
is to convert the visibility map to a new format, does not suit the
existing design anyway.  So just rip it out.

Author: Masahiko Sawada, slightly revised by me.
Discussion: 20160215211313.GB31273@momjian.us