Ignore XLOG archive check when creating restart point on mirror segment
When a mirror is running with archive_mode=on, it may incorrectly mark
WAL segments as ready to be archived (".ready" file in the
$datadir/pg_xlog/archive_status directory) when checking to see if any
WAL segments can be recycled. This causes the marked WAL segment to
never be recycled. It can also cause issues when the mirror is
promoted and tries to start archiving WAL since the old stale WAL
segment may already exist in the archive location.

To fix the issue, ignore the XLOG archive check when creating a
restart point on a mirror segment. We really shouldn't be doing
anything relating to WAL archiving on the mirror segment while it's in
standby mode anyways since continuous WAL archiving for
mirrors/standbys is not supported until its introduction in Postgres
9.5.

This patch was heavily inspired from upstream postgres commit:
https://github.com/postgres/postgres/commit/78ea8b5daab9237fd42d7a8a836c1c451765499f
```
Fix WAL recycling on standbys depending on archive_mode

A restart point or a checkpoint recycling WAL segments treats segments
marked with neither ".done" (archiving is done) or ".ready" (segment is
ready to be archived) in archive_status the same way for archive_mode
being "on" or "always".  While for a primary this is fine, a standby
running a restart point with archive_mode = on would try to mark such a
segment as ready for archiving, which is something that will never
happen except after the standby is promoted.

Note that this problem applies only to WAL segments coming from the
local pg_wal the first time archive recovery is run.  Segments part of a
self-contained base backup are the most common case where this could
happen, however even in this case normally the .done markers would be
most likely part of the backup.  Segments recovered from an archive are
marked as .ready or .done by the startup process, and segments finished
streaming are marked as such by the WAL receiver, so they are handled
already.

Reported-by: Haruka Takatsuka
Author: Michael Paquier
Discussion: https://postgr.es/m/15402-a453c90ed4cf88b2@postgresql.org
Backpatch-through: 9.5, where archive_mode = always has been added.
```

Upstream postgres commit reference for continuous WAL archiving
support for standbys:
https://github.com/postgres/postgres/commit/ffd37740ee6fcd434416ec0c5461f7040e0a11de