Add valid Default Partition check in ATPExecPartSplit(). (#14188)
In GP6's ALTER TABLE SPLIT PARTITION INTO statements, when the INTO
cause contains a default partition which is not defined, it would cause
a PANIC, which is caused by a string copy operation on the NULL partition
definition. But in GP7, this wrong syntax statement would be intercepted by
the parser stage. By this commit, I added check whether the default partition
inside the INTO cause is valid, and abstracted a method to remove redundant
code performing the similar work.