Free underlying Dynamic Seq Scan structures after processing of each partition (#14001)
The tables with high count of partitions can drain high amount of executor
memory for processing. This can be especially critical for column-oriented
tables, because we allocating a buffer (which is triple of block size) for each
column. Following the standard flow, we free such buffers at executor end. But
it seems, we can safely free some structures, including these buffers, after
processing of each partition. To do this in the uniform way, new *_afterscan()
access methods introduced by this patch.

Related to #13725.