Generate correct result for query on date and timestamp predicate in partition table scenario
ORCA generates wrong result for a query on partitioned table, in a scenario where the query
has different type of predicates at the same time like date and timestamp.

Eg: select max(datedday) from table_rangePartition where
datedday='2022-10-23' or datedday=('2022-10-23'::date -interval '1 day');

In a query with multiple predicates of different kind, during partition elimination a function
is selected to read the predicate values. This function selection was done “once” for the whole query.

Selecting the function once itself was the root cause of the failure, as in the case of
query with different type of predicates, using the same function to access different type of
predicate caused generation of garbage values. As in the query above, the same function to access
timestamp was used to access date type predicate, which generated wrong results.

To fix, the problem, the retention of old functions was removed in static partition elimination
for range and list partition.