icproxy isolation2 test fix
Commit 5d977f6 introduced a change to recoverseg querying
gp_stat_replication which requires motion on the segments.
This caused a failure on 6X pipeline for ic_proxy isolation2 test

recoverseg test changes dbid to a new dbid which doesn't have an entry
in icproxy addresses. This commit fixes the test by doing an icproxy
address reset after the dbid is changed to pick up new dbids.